Rename module and add readme
