+++ module/localisation/BallFilter/data/config/BallFilter.yaml
@@ -1,0 +1,5 @@
+# Controls the minimum log level that NUClear log will display
+log_level: DEBUG
+
+# Smoothing factor used in exponential filter on ball position
+smoothing_factor: 0.2
