--- module/localisation/VisionBallLocalisation/README.md
@@ -1,13 +1,0 @@
-VisionBallLocalisation
-================
-
-## Description
-
-
-## Usage
-
-
-## Emits
-
-
-## Dependencies
