--- module/localisation/VisionBallLocalisation/data/config/VisionBallLocalisation.yaml
@@ -1,4 +1,0 @@
-# Controls the minimum log level that NUClear log will display
-log_level: WARN
-
-smoothing_factor: 0.1
